Avformat demuxer: kill 2 warnings about signedness

+++ modules/demux/avformat/demux.c
@@ -638,7 +638,7 @@ static block_t *BuildSsaFrame( const AVPacket *p_pkt, unsigned i_order )
         return NULL;
 
     char buffer[256];
-    const size_t i_buffer_size = __MIN( sizeof(buffer) - 1, p_pkt->size );
+    const size_t i_buffer_size = __MIN( (int)sizeof(buffer) - 1, p_pkt->size );
     memcpy( buffer, p_pkt->data, i_buffer_size );
     buffer[i_buffer_size] = '\0';
 
@@ -650,7 +650,7 @@ static block_t *BuildSsaFrame( const AVPacket *p_pkt, unsigned i_order )
     if( sscanf( buffer, "Dialogue: %d,%d:%d:%d.%d,%d:%d:%d.%d,%n", &i_layer,
                 &h0, &m0, &s0, &c0, &h1, &m1, &s1, &c1, &i_position ) < 9 )
         return NULL;
-    if( i_position <= 0 || i_position >= i_buffer_size )
+    if( i_position <= 0 || (unsigned)i_position >= i_buffer_size )
         return NULL;
 
     char *p;
